requirements.txt
Django
pytz
Whoosh
django-localflavor
django-auth-ldap
feedparser
vobject

django-haystack
django-guardian

django-dept-classes
django-dept-directory
django-dept-gradstudents
django-dept-job-postings
django-dept-pages
django-dept-personpages
django-dept-publications
django-dept-seminars
django-dept-swingtime
django-shouts
django-uofm
django-dept-exams
django-dept-course-groups
django-fileshare
django-face-detect
django-dept-committee
django-dept-mgmt
django-markuphelpers
django-admin-export
django-dept-courseplanning

django-science-templates
